Updated patients - BigDecimal instead of INT for pesel and telNum

+++ src/main/java/com/kodilla/clinic/dto/PatientDto.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -17,14 +18,23 @@ public class PatientDto {
     private String surname;
     private String address;
     private LocalDate birthDate;
-    private int pesel;
-    private int telNum;
+    private BigDecimal pesel;
+    private BigDecimal telNum;
     private String email;
     private boolean inUrgency;
     private List<Integer> appointmentsIds;
     private List<Integer> evaluationsIds;
 
-    public PatientDto(Integer patient_id, String name, String surname, String address, int pesel, int telNum, String email, boolean inUrgency, List<Integer> appointmentsIds, List<Integer> evaluationsIds) {
+    public PatientDto(Integer patient_id,
+                      String name,
+                      String surname,
+                      String address,
+                      BigDecimal pesel,
+                      BigDecimal telNum,
+                      String email,
+                      boolean inUrgency,
+                      List<Integer> appointmentsIds,
+                      List<Integer> evaluationsIds) {
         this.patient_id = patient_id;
         this.name = name;
         this.surname = surname;
